requirements.txt
numpy
pandas
scikit-learn
xgboost
joblib
Pillow
selenium
beautifulsoup4
requests
webdriver-manager
janome
janome
scikit-learn
pandas
